requirements.txt
beautifulsoup4==4.11.2
black==23.1.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
idna==3.4
loguru==0.6.0
mypy-extensions==1.0.0
numpy==1.24.1
openpyxl==3.1.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
progressbar==2.5
pyyaml-tags==1.0.4
requests==2.28.2
soupsieve==2.3.2.post1
tqdm==4.64.1
urllib3==1.26.14

